requirements.txt
Django==5.2
djangorestframework==3.16.1
pydantic==2.12.5
python-dotenv==1.2.2
ollama==0.1.6
